+++ src/components/HomePage/HomepageSinglefeed.js
@@ -69,7 +69,7 @@ const HomepageSinglefeed = () => {
                     <textarea
                         id="message"
                         rows="4"
-                        class="ml-2 block p-2.5 h-16 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-gray-300 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                        class="ml-2 block p-2.5 h-16 w-full text-sm text-gray-900 bg-gray-50 rounded-lg border border-emerald-700 focus:ring-emerald-700 focus:border-emerald-700 "
                         placeholder="Write something here..."></textarea>
                     <div className='absolute right-2 top-10'>
                         <FaFileVideo className='cursor-pointer' />
